parking-forecast/src/requirements.txt
pyyaml==6.0.2
pandas==1.3.4
numpy==1.21.1
tensorflow==2.4.3